+++ src/modules/ipfix/aggregator/FlowHashtable.h
@@ -31,7 +31,7 @@ class FlowHashtable : public BaseHashtable
 {
 public:
 	FlowHashtable(Source<IpfixRecord*>* recordsource, Rule* rule,
-			uint16_t minBufferTime, uint16_t maxBufferTime, uint8_t hashbits);
+			uint32_t minBufferTime, uint32_t maxBufferTime, uint8_t hashbits);
 	virtual ~FlowHashtable();
 
 	void aggregateDataRecord(IpfixDataRecord* record);
