+++ src/modules/ipfix/aggregator/PacketAggregator.cpp
@@ -69,8 +69,8 @@ void PacketAggregator::receive(Packet* e)
 /**
  * creates hashtable for this aggregator
  */
-BaseHashtable* PacketAggregator::createHashtable(Rule* rule, uint16_t minBufferTime,
-		uint16_t maxBufferTime, uint8_t hashbits)
+BaseHashtable* PacketAggregator::createHashtable(Rule* rule, uint32_t minBufferTime,
+		uint32_t maxBufferTime, uint8_t hashbits)
 {
 	return new PacketHashtable(this, rule, minBufferTime, maxBufferTime, hashbits);
 }
